Add schema manager (#23)
* Add schema manager

* Actualize

* Remove usage of SQLite column

* Make console optional in composer

* Add RbacDbInit command to example [skip ci]

* Remove unnecessary line [skip ci]

* Update example, remove command [skip ci]

* Add base code to schema manager

* Replace command with schema manager in tests

* Add conditions for has table

* Dump SQL

* Apply fixes from StyleCI

* Revert + remove force

* Add sqlite migrations [skip ci]

* Add sqlite sql [skip ci]

* Add pgsql migrations [skip ci]

* Add pgsql sql [skip ci]

* Add mysql migrations [skip ci]

* Add mysql migrations 2 [skip ci]

* Add mssql migrations [skip ci]

* Add oracle migrations [skip ci]

* Add oracle and mssql sql [skip ci]

* Update docs [skip ci]

* Update docs 2 [skip ci]

* Improve coverage

* Fix mutants

* test for schema manager, create all multiple - perform additional assert / check

* Update changelog and composer suggest

* Fix namespace in example [skip ci]

* Update README with the structure of SQL files [skip ci]

* Add missing PHPDoc

* [ci-review] Apply changes from Rector action.

* Apply fixes from StyleCI

* Actualize after latest discussions

* Apply fixes from StyleCI

* Update CHANGELOG.md


* Update README.md


* Update src/DbSchemaManager.php


* Update tests/Sqlite/SchemaManagerTest.php


* Apply fixes from StyleCI

* Update tests/Pgsql/SchemaManagerTest.php


* Update tests/Oracle/SchemaManagerTest.php


* Apply fixes from StyleCI

* Update tests/Mysql/SchemaManagerTest.php


* Apply fixes from StyleCI

* Update tests/Mssql/SchemaManagerTest.php


* Apply fixes from StyleCI

* Update tests/Base/SchemaManagerTest.php


* Apply fixes from StyleCI

* Rename all tests at once

* Update README.md


* Update README.md


* Update README.md


* Remove ambiguous sentence [skip ci]

---------
