+++ composer.json
@@ -21,7 +21,6 @@
     "require": {
         "ext-pdo": "*",
         "php": "^8.0",
-        "symfony/console": "^5.3|^6.0",
         "yiisoft/db": "^1.0",
         "yiisoft/rbac": "^1.0.2"
     },
@@ -41,7 +40,8 @@
         "yiisoft/db-mysql": "For using with MySQL",
         "yiisoft/db-pgsql": "For using with PosgtgreSQL",
         "yiisoft/db-mssql": "For using with Microsoft SQL Server",
-        "yiisoft/db-oracle": "For using with Oracle"
+        "yiisoft/db-oracle": "For using with Oracle",
+        "yiisoft/yii-db-migration": "For automating schema migration"
     },
     "autoload": {
         "psr-4": {
