+++ sql/oracle-down.sql
@@ -1,0 +1,3 @@
+DROP TABLE "auth_item_child";
+DROP TABLE "auth_assignment";
+DROP TABLE "auth_item";
