--- tests/Mssql/Command/RbacDbInitTest.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Rbac\Db\Tests\Mssql\Command;
-
-use Yiisoft\Rbac\Db\Tests\Mssql\MssqlTrait;
-
-class RbacDbInitTest extends \Yiisoft\Rbac\Db\Tests\Base\Command\RbacDbInitTest
-{
-    use MssqlTrait;
-}
