+++ tests/Pgsql/DbSchemaManagerTest.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Rbac\Db\Tests\Pgsql;
+
+class DbSchemaManagerTest extends \Yiisoft\Rbac\Db\Tests\Base\DbSchemaManagerTest
+{
+    use PgsqlTrait;
+}
